tunniplaan, õppeaste -> õppetase

+++ hois_back/src/main/java/ee/hitsa/ois/web/GeneralMessageController.java
@@ -20,6 +20,7 @@ import org.springframework.web.bind.annotation.RestController;
 import ee.hitsa.ois.domain.GeneralMessage;
 import ee.hitsa.ois.service.GeneralMessageService;
 import ee.hitsa.ois.service.security.HoisUserDetails;
+import ee.hitsa.ois.util.AssertionFailedException;
 import ee.hitsa.ois.util.HttpUtil;
 import ee.hitsa.ois.util.UserUtil;
 import ee.hitsa.ois.util.WithEntity;
@@ -42,29 +43,38 @@ public class GeneralMessageController {
 
     @GetMapping
     public Page<GeneralMessageDto> search(HoisUserDetails user, @Valid GeneralMessageSearchCommand criteria, Pageable pageable) {
+        assertSchoolAdmin(user);
         return generalMessageService.search(user.getSchoolId(), criteria, pageable);
     }
 
     @GetMapping("/{id:\\d+}")
     public GeneralMessageDto getGeneralMessage(HoisUserDetails user, @WithEntity("id") GeneralMessage generalMessage) {
+        assertSchoolAdmin(user);
         UserUtil.assertSameSchool(user, generalMessage.getSchool());
         return GeneralMessageDto.of(generalMessage);
     }
 
     @PostMapping
     public ResponseEntity<Map<String, ?>> createGeneralMessage(HoisUserDetails user, @Valid @RequestBody GeneralMessageForm form) {
+        assertSchoolAdmin(user);
         return HttpUtil.created(generalMessageService.create(user, form));
     }
 
     @PutMapping("/{id:\\d+}")
     public GeneralMessageDto updateGeneralMessage(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestBody = true) GeneralMessage generalMessage, @Valid @RequestBody GeneralMessageForm form) {
+        assertSchoolAdmin(user);
         UserUtil.assertSameSchool(user, generalMessage.getSchool());
         return getGeneralMessage(user, generalMessageService.save(generalMessage, form));
     }
 
     @DeleteMapping("/{id:\\d+}")
     public void deleteGeneralMessage(HoisUserDetails user, @WithVersionedEntity(value = "id", versionRequestParam = "version") GeneralMessage generalMessage, @SuppressWarnings("unused") @RequestParam("version") Long version) {
+        assertSchoolAdmin(user);
         UserUtil.assertSameSchool(user, generalMessage.getSchool());
         generalMessageService.delete(generalMessage);
+    }
+
+    private static void assertSchoolAdmin(HoisUserDetails user) {
+        AssertionFailedException.throwIf(!user.isSchoolAdmin(), "No rights");
     }
 }
